// frontend/Pixion.LearnRag.ReactClient/src/api/model/answerPromptTemplate.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Learn RAG API
 * OpenAPI spec version: v1
 */

export type AnswerPromptTemplate = {
  /** @nullable */
  context?: string | null;
  /** @nullable */
  question?: string | null;
  /** @nullable */
  readonly templateString?: string | null;
};